Add chassis column to MAC_Binding table
The new chassis column in MAC_Binding should
act as an "owner" order of that particular row.
This can be utilized by MAC binding aging mechanism,
where only single is responsible for that particular
record. In order to keep it consistent the chassis
that owns the MAC binding is the one that created it,
with one exception, when the MAC of the binding has
changed the ownership is transferred to the controller
that made this update. MAC binding rows without
chassis are removed by northd.

Reported-at: https://bugzilla.redhat.com/2084668
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>