Add MAC binding aging mechanism
Add MAC binding aging mechanism that utilizes
the ownership of MAC binding row. The controller
that "owns" the MAC binding will track idle_age
statistics for related OpenFlows (table 66 and 67).

If the idle_age exceeds the threshold value for aging
the MAC binding row is removed. The threshold is set to
60 secs. The removal won't happen exactly at the 60
second mark, but can happen any time later on, depending
on timing of the ovn-controller loop.

Reported-at: https://bugzilla.redhat.com/2084668
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>