OVN-CI: Add test cases with monitor-all enabled.
currently ovn ci only has one test case with the option
ovn-monitor-all enabled, this patch will add one more
execution with option ovn-monitor-all=true for each test case that
are wrapped by OVN_FOR_EACH_NORTHD macro.

This will more or less double the number of test cases.
It is possible to select a reduce set of test cases using -k "keywords".
Keyword such as
        dp-groups=yes
        dp-groups=no
        parallelization=yes
        parallelization=no
        ovn-northd
        ovn-northd-ddlog
        ovn_monitor_all=yes
can be used to select a range of tests, as title is searched as well.

For instance, to run ovn-monitor-all tests, with dp-groups enabled and ddlog disabled:
        make check TESTSUITEFLAGS="-k dp-groups=yes,ovn_monitor_all=yes,\!ovn-northd-ddlog"

Signed-off-by: Mohammad Heib <mheib@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>