Always funnel multichassis port traffic through tunnels
Before the patch, for switches with a localnet port,

- traffic to a multichassis port was funneled through tunnels; but
- traffic from a multichassis port was sent through localnet.

This is not optimal because:
- bidirectional sessions are sent through two separate paths;
- leaking multichassis traffic to physical makes upstream switch flip
  port's mac address location.

One of the effects of this behavior is prolonged network downtime when
live migrating a VM attached to a physical network (compared to
downtime observed for pure geneve networks).

As per my unscientific local testing, this patch reduces network
downtime during OpenStack VM live migration from 1.0s+ to 0.05-0.10s,
on par with geneve networks.

Fixes: 10398c1f51d5 ("Clone packets to all port chassis")
Signed-off-by: Ihar Hrachyshka <ihrachys@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>