Allow arbitrary args to be passed to called binary
It is common to pass ovn-ctl options via an /etc/sysconfig file.
It would be useful to be able to pass custom --remote options or
additional DBs to listen to via this file. This would give end
users the ability to have more fine-grained control without
having to modify ovn-ctl.

Signed-off-by: Terry Wilson <twilson@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>