northd: Add feature to log reply and related ACL traffic.
It can be desirable for replies to stateful ACLs to be logged. And in
some cases, it can actually be a bit confusing why they aren't logged.
Consider a situation where a port group called "port_group" exists and
logical switch ports swp1 and swp2 belong to it. We create the following
ACL, where logging is enabled:

from-lport 100 'inport == @port_group' allow-stateless

swp1 sends traffic to swp2 and swp2 responds within the same connection.
In this case, the logs will show both the packets from swp1 to swp2, as
well as the response packets from swp2 to swp1. This is because the
traffic matches the ACL in both directions.

Now change the ACL:

from-lport 100 'inport == @port_group' allow-related

Now with the same traffic pattern, the packets from swp1 to swp2 are
logged, but the reply packets from swp2 to swp1 are not. Why is that?

The reason is that as a shortcut, when stateful ACLs are used, reply
traffic bypasses ACL evaluation. Therefore, even though it may appear
that the reply should match the ACL, the ACL is never even evaluated,
and therefore is not logged.

With this change, reply traffic still bypasses ACL evaluation, but the
reply can still be logged.

Since logging reply traffic requires adding more flows, it is not
enabled by default. In order to have reply traffic logged, the ACL
must have logging enabled, be stateful, have a label, and have the new
log-related option set to true.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2031150

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Reviewed-by: Frode Nordahl <frode.nordahl@canonical.com>
Acked-by: Numan Siddique <numans@ovn.org>
Acked-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>