controller: Fix wrong controller action definition
The controller action had a wrong wrong userdata_len due to
missing substraction of previous data len. The pointer to oc
was not moved correctly so it could in theory point to wrong
memory. In order to fix that expose encode_start_controller_op
and encode_finish_controller_op which are helper methods
to define the controller action. Use those instead of
manually creating the action. At the same time use stack
buffer for the ofpbuf which should be large enough to hold
related data without any heap allocations.

Fixes: e52c2451 ("physical.c: Fix bug of wrong use in vm migration")
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>