northd: Always propagate LB proto to lflows
The current behavior is that load balancers
default to tcp protocol if it is not specified.
However, this is not propagated to lflows which
specify proto only when port is specified as part
of the protocol e.g. udp.port == X. Always include
the protocol in the lb flow even if the port is not
specified.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>