northd: Retrieve load balancer options only once.
Options like 'add_route' and 'neighbor_responder' are looked up
in the smap for each load balancer for each datapath and it takes
significant amount of time.  Checking them only once to speed up
processing.  'skip_snat' and 'event' are not that critical, but
it's better to get all of them in a single place.

Also using enum for the neighbor responder mode to avoid costly
string comparison on a hot path.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>