pinctrl: Do not send garp if vif is not on this chassis
whem vm migrating from node A to node B. One ovs interface is
created on B , and attached to br-int. At this time, vm is
still on node A. On node B, In procedure of pinctrl_run,
function send_grap_rarp_prepare notice the created interface,
then send a grap, which will lead the mac table on network
changed. Then, network flows to vm on node A will be forwards
to node B. But the process of vm migration has not accomplished.
So, other clients lost connection to vm.

Here, if port's chassis is not on node B, grap packet sent
from controller is stopped, fix the bug descriped upper.

Signed-off-by: wangchuanlei <wangchuanlei@inspur.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>