northd: don't include disabled LSP's IP to Load Balancing
If one has a UDP load balancer with backend IP which is located under
disabled LSP, such backend would be threated as alive and marked as
'online' on Service_Monitor table and added to load balancing as well.
Though such LSP can't receive any traffic, this load balancer will be
broken by mentioned behaviour.

This patch resolves this issue for Load Balancers with enabled health
check: if LSP is disabled, it wont be added to Service_Monitor and to
Load Balancing flow.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>