northd: Use bitmaps instead of exclusive hash map for datapath groups.
Every time a new logical flow is created, ovn-northd creates an
exclusive hash map to collect all the datapaths this logical flow is
applicable to.  While adding a new datapath to existing logical flow
hmapx will perform a lookup by hash and then add a new entry by
allocating a piece of memory and putting it into a hash map.  When
all the logical flows are created and all the datapath groups are
ready, northd is collecting all the unique groups by comparing them
to each other, destroying duplicates and replacing them with a pointer
to a one true group.  After that, northd is analyzing all the existing
flows in the Sb DB and comparing their groups to ones that the flow
should have.  At the end all the remaining exclusive hash maps
containing datapath groups are destroyed.

The process is fairly inefficient.  The number of unique datapath
groups is typically much smaller than the number of logical flows.
So, in high-scale environments we may end up generating hundreds
of thousands equal groups just to discard them later.
Every element in a group is a separately allocated piece of memory
added to a hash map.  This imposes significant cost of allocating
them and freeing later.  Comparison of 2 hash maps to check for
equality involves looking up every element of one hash map in another,
which is not a very fast operation either.  Also hash maps are
inherently cache-unfriendly, so every access during iteration
likely results in a cache miss.

As a result, all these datapath group manipulations end up taking
most of the processing time in northd.

Let's enumerate all the logical datapaths we have assigning a
unique index from zero to the total number of datapaths.  And
create an array of datapaths for a quick access by their index.
Having that, instead of collecting datapath pointers in hash
maps, we can use bitmaps instead and turn on bits that correspond
to datapaths relevant to particular group.  Later we can reconstruct
the group by iterating over a bitmap and taking required datapaths
from the array.

Advantages:

 - Bitmaps are relatively small.  It will be only 125 bytes to
   accommodate a setup with 1000 datapaths.  At the same time hmapx
   with just one element will end up with 2 separate memory allocations
   to allocate at least 4 buckets and a hmapx_node.  In total that can
   be less than a bitmap with only one datapath, but it will be only
   2-3x less and comparison becomes irrelevant as soon as there is at
   least one big datapath group.

 - Bitmaps are trivial to compare.  Simple memcmp does the trick.
   And memcmp will have one the most optimized implementations for
   the actual hardware in use.

 - Flipping bits or checking if a certain bit is set in constant time.
   Asymptotically same as in hash maps, but the constant is much lower.

 - Cache-friendly.  Depending on a system and the scale, the whole map
   can fit into a single cache line or two.  All the data is sequential
   in memory.

 - Reduced number of memory allocations/deallocations.  Bitmap is
   allocated and freed only once as a single piece.  hmapx allocates
   memory for every datapath.

Disadvantage:

 - Counting number of bits in a bitmap is a bit heavier than just
   getting hmapx_count(), but we're not doing that very frequently,
   so it doesn't seem to impact performance in any meaningful way.

Testing with ovn-heater shows up to 3x performance improvement for
500-node density-heavy scenario.  In this case northd poll intervals
went down from 10 to 3 seconds, driving the test maximum latency from
21 down to 6.5 seconds.  And overall test time went down from 2400 to
720 seconds.  Memory consumption for the northd process also reduced
from 3.9 GB to 1.6 GB RSS.

One design consideration was to make 'datapaths_array' and the
'n_dataptahs' variables global.  Logically, from the perspective of
the I-P engine, they should be part of the northd-generated 'data'.
And it's not a problem for the array itself, since it is only used in
a couple of places.  However, 'n_datapaths' is used practically
everywhere, including ovn_lflow_add(), so we'll have to pass it around
in most of the functions in northd as an argument.  Taking that into
account, I decided to make them both global to avoid touching half of
the northd code.
We also have a couple of other global variables already (mainly for
parallel processing) that should technically be part of the I-P data,
but it is way simpler to keep them global.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>