northd: drop traffic to disabled LSPs in ingress pipeline
Prior to this patch traffic to LSPs, which are disabled with
`ovn-nbctl lsp-set-enabled <LSP> disabled` was dropped in the end of
lswitch egress pipeline.  This means that traffic is processed in vain:
- traffic, which should be dropped, first travels from one chassis to
  another (if source/dest LSPs reside on different nodes) and dropped on
  the destination chassis;
- when such traffic reaches destination chassis, if stateful services are
  enabled within logical switch, first traffic is sent to conntrack and
  is dropped after that.

So it is costly to drop traffic in such manner especially in case LSP is
disabled to prevent chassis and/or VM attack by any harmful traffic.

This patch changes "to-lport" drop behaviour.  Now it is dropped in
lswitch ingress pipeline to avoid sending traffic to disabled LSP from
one chassis to another.
Traffic doesn't reach conntrack in destination LSP's zone now as well.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>