actions: introduce commit_lb_aff action
commit_lb_aff action translates to an openflow "learn" action that
inserts a new flow in the OFTABLE_CHK_LB_AFFINITY table. The new flow is
used to match on the the 5-tuple and set REGBIT_KNOWN_LB_SESSION bit.
Moreover the new flow stores backend IP and port in register REG4 and
REG8[0..15] respectively.

Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>