ovn-controller: Fix some issues with CT zone assignment.
An issue was filed where CT zone 0 was assigned to both a logical router
SNAT and to a logical port. CT zone 0 is typically "reserved" and not
assigned by ovn-controller; however, since SNAT zones are configurable,
it is possible for ovn-controller to assign this zone at the CMS's
request. This accounts for how CT zone 0 can be assigned for SNAT. There
was also a small bug in the incremental processing that could result in
a logical port being assigned zone 0.

In the specific issue report, CT zones were restored from the OVSDB when
ovn-controller started, and the conflicting CT zones were already
present. ovn-controller dutifully loaded these zones up. But then there
was nothing that would allow for the conflict to be resolved afterwards.

It is unknown how these conflicts entered into the OVSDB in the first
place. This change does not purport to prevent conflicts from entering
the OVSDB. However, it does make the following changes that should
further safeguard against unwanted behavior:

* ct_zones_runtime_data_handler() now assigns zones starting at
  1 instead of 0. This makes it use the same range as update_ct_zones().
* update_ct_zones() now keeps a new simap for zones that are assigned
  but not due to an SNAT zone request. This allows for us to guarantee
  that when there is a conflict between a previously auto-assigned CT
  zone and a newly-requested zone, we are guaranteed to remove the
  auto-assigned CT zone.
* The removal of conflicting auto-assigned CT zones is now performed
  before dealing with the newly requested zone. This makes it so that if
  we load conflicting zones from the OVSDB or if there is some issue
  that results in conflicting zones being assigned, we will correct the
  issue.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2126406
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Co-authored-by: Ales Musil <amusil@redhat.com>
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>