actions: limit possible OF group bucket count
It is possible to send OpenFlow group_mod message to OVS to create a
group with any number of buckets:

ovs-ofctl dump-groups br-int
NXST_GROUP_DESC reply (xid=0x2):
 group_id=4,type=select,selection_method=dp_hash,bucket=bucket_id:0,weight:100,actions=load:0x1->OXM_OF_PKT_REG4[48..63],resubmit(,20),...bucket=bucket_id:10000,...

This patch introduces a limit of buckets that may be requested to 1024.
In case the limit is reached, ovn-controller will write WARN log about
this fact.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>