ovn-controller: Only set monitor conditions on available tables.
Setting the monitor condition on a table that's not part of the SB
schema triggers an error from the server side, a downgrade to monitor v0
(from v2) and also causes the conditional monitoring sequence numbers to
go out of sync.

Use the <DB>rec_server_has_<TABLE>_table() API to determine if a table
is present in the server schema and therefore usable in the monitor
condition.

This commit also bumps the OVS submodule because the following fix is
needed:
  a787fbbf9dd6 ("ovsdb-cs: Consider default conditions implicitly acked.")

We also pick up the following relevant OVS changes:
  b8bf410a5c94 db-ctl-base: Use partial map/set updates for last add/set commands.
  55b9507e6824 ovsdb-idl: Add the support to specify the uuid for row insert.

Reported-by: Numan Siddique <numans@ovn.org>
Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2151063
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>