.ci: ovn-kubernetes: Skip session affinity related tests.
The OVN feature used to implement session affinity timeout doesn't exist
on this branch.  ovn-kubernetes code already moved to implementing the
feature by using the newer OVN option so, in order to keep testing as
many features as possible in CI, we now disable all session affinity
tests (but keep executing the rest).

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>