ovn-controller: Fix initial requested SNAT zone assignment.
update_ct_zones() attempts to give preference to requested SNAT zones
over auto-assigned CT zones. However, there is a subtle bug. Requested
SNAT zones are only assigned if there is an auto-assigned CT zone to
replace. If a logical router is created, and its requested SNAT zone is
assigned in the same transaction, then that logical router will not yet
have an auto-assigned CT zone on it. Therefore, the requested SNAT zone
has no CT zone to replace, and so the logical router is auto-assigned a
zone.

This bug is somewhat hidden. The next time a recompute is needed in
ovn-controller, then the requested SNAT zone will replace the
auto-assigned zone. This is why none of our tests caught this issue. Our
tests always created the logical router in one transaction, then set the
requested SNAT zone later. The initial transaction would auto-assign an
SNAT zone for the logical router, and when we set the requested zone
later, a recompute would properly replace the auto-assigned zone with
the requested zone.

This change fixes the issue by setting the requested zone no matter if
the logical router already has an auto-assigned zone. A simple test has
been added to ensure that the behavior works as expected.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2160403
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>