lflow: Warn about always false matches.
Logical flow matches that always evaluate to false should never be
programmed by ovn-northd (or as ACLs).  If that's the case, log a
warning.

Note: matches that cannot be parsed because they refer to empty
or non existent port_groups/address_sets/templates don't generate
a valid expression structure so they trigger a different warning.
This behavior was already present before the current commit.

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>