northd: Add thread safety analysis to lflow hash locks.
Even though it's not possible to enable a full thread safety analysis
due to conditional mutex taking and unpredictable order [1], we can add
most of the functionality with a fake mutex.  It can detect nesting
of hash locks, incorrect locking sequences and some other issues.
More details are given in a form of a comment in the code to give this
fake mutex more visibility.

The fake mutex is declared as extern to not actually use any memory
or trigger 'unused' warnings.

No variables or structure fields are marked as GUARDED, because we
still need a lockless access once the parallel part is over.

[1] https://clang.llvm.org/docs/ThreadSafetyAnalysis.html#no-conditionally-held-locks

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>