northd: Use bitmaps for LB lists of switches and routers.
Lists of switches and routers to which a certain load balancer is
applied are frequently iterated in order to add all these datapaths
into datapath group bitmap.  They would be easier to operate with,
if they were bitmaps themselves in the first place.

This change also lays out the infrastructure for more elegant creation
of logical flows with datapath groups in the next commits.

Acked-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>