northd: Create metered flows with dp groups if CoPP is not configured.
Previous change introduced the ovn_lflow_add_with_dp_group() wrapper
and applied it to all the non-metered logical flows.  Metered ones are
a bit more complex, but still can be optimized to use the new wrapper
in cases where CoPP is not enabled or enabled only on some datapaths.

Iterating over all the datapaths and excluding metered ones from the
bitmap.  For the remaining datapaths the new wrapper can be used.
Metered flows can be created right away.  This might be slightly
slower for a case where CoPP is enabled due to more hash calculations,
but that can be optimized later, if necessary.

With this change, there are no more users of ovn_lflow_add_at_with_hash(),
so it is inlined into ovn_lflow_add_at(). 'struct lrouter_nat_lb_flow'
degraded to just an action string, so inlined into ctx structure.
Initializer is no longer needed.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>