northd: Use LB port_str in northd.
As mentioned in the structure definition in lb.h, northd should never
use 'vip_port' instead it should always use 'port_str'.  That's because
it might not always be possible to evaluate the port value in northd,
e.g., when using LB templates.

This commit also updates other occurrences of 'vip_port' replacing them
with 'port_str' in northd.  Those were not causing issues because they
were only in cases when templates are not supported (LB affinity for
example) but it's better to have consistency in style.

Fixes: 7b94d212f694 ("northd: Refactor build_lrouter_nat_flows_for_lb function")
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>