tests: Decrease the number of zones and switches for interconnection
When we run the tests with address sanitizers enabled
the interconnection test consumes a lot of memory,
around 5 GB (on my env) per permutation. To prevent
any OOM kills on Github CI reduce the number of
zones and transit switches to 3 as this lowers
the memory consumption to around 2 GB per permutation.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>