expr: Combine OR sub-expressions into supersets.
If two simple CMP expressions within OR have exact same masks, and
their values differ in a single bit, these expressions can be combined
into one by removing that bit from the mask.

E.g. Expression (tcp.src == 6441 || tcp.src == 6443) can be
     simplified down to single tcp.src == 0x1929/0xfffd.

     6441 = 0x1929 = 0b0001100100101001
     6643 = 0x192b = 0b0001100100101011
                                     ^
            0x1929   0b0001100100101001
               /   =          /
            0xfffd   0b1111111111111101

Since the bit can be either 0 or 1, there is no need to check it at all.

Add checking of possibility to combine two expressions into a new
superset while handling supersets in crushed OR expressions.  This
allows to combine entries multiple times as well as remove entries
that are subsets of newly created supersets.

The loop will not spin for way too long, because every jump back means
removal of at least one existing entry from the list.

Change allows to efficiently squash IP address sets as well as matches
on other fully maskable fields like L4 ports and MAC addresses.  E.g.:

 $ ./tests/ovstest test-ovn expr-to-flows <<< \
     "ip4.src == {172.168.12.0/24, 172.168.13.0/24, 172.168.15.0/24}"
 ip,nw_src=172.168.15.0/24
 ip,nw_src=172.168.12.0/23

 $ ./tests/ovstest test-ovn expr-to-flows <<< \
     "ip4.src == {172.168.12.0/24, 172.168.13.0/24, 172.168.14.0/24, \
                  172.168.15.0/24}"
 ip,nw_src=172.168.12.0/22

 In this case, first we combine 12 and 13, then 14 and 15, then results
 of two previous combinations.

It's a greedy algorithm, i.e. it makes locally optimal decisions on
each step that may or may not be globally optimal:

  $ ./tests/ovstest test-ovn expr-to-flows <<< \
        "ip4.src == {172.168.12.0/24, 172.168.13.0/24, 172.168.14.0/24, \
                     172.168.15.0/24, 172.169.12.0/24}"
  ip,nw_src=172.168.12.0/22
  ip,nw_src=172.169.12.0/24

  But

  $ ./tests/ovstest test-ovn expr-to-flows <<< \
        "ip4.src == {172.169.12.0/24, 172.168.12.0/24, 172.168.13.0/24, \
                     172.168.14.0/24, 172.168.15.0/24}"
  ip,nw_src=172.168.13.0/255.255.253.0
  ip,nw_src=172.168.14.0/24
  ip,nw_src=172.168.12.0/255.254.255.0

In the case above it would be better to combine all the 172.168.*
subnets, but it doesn't happen if 172.169.12.0/24 is at the top of
the list, because it can be paired with 172.168.12.0/24 instead.
However, produced results are good enough in most cases.  Full
non-greedy solution will likely be much more computationally complex.

Tests updated to reflect this change.  Extra couple of specific tests
added.  Tests that depend on the exact number of OpenFlow rules
updated to use addresses that cannot be combined.

This change makes address set I-P less efficient, but it also
reduces the number of generated OpenFlow rules, so very efficient
I-P is less important.

Signed-off-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>