northd: Split switch and router datapaths.
Similar to the previous commit that splits switch ports and router
ports, this patch splits the two types of datapaths in different hmap
tables.  This is also mainly to help for future incremental processing
implementation. As a byproduct, the recompute performance is improved
slightly. (Together with the previous commit that splits ports, scale
test shows around ~4% improvment)

Note: This patch didn't change the ovn_datapath structure, which is
used by both switches and routers. It may be a followup improvement in
the future if necessary, e.g. to reduce memory footprint.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>