northd: Move the datapaths arrays to ovn_datapaths struct.
In the current code, many functions depends on the global datapaths
arrays variables, which is convenient. However, for incremental
processing framework to track dependencies properly, we should avoid
using global variables and purely depends on engine inputs (and function
inputs, for the functions called by engine nodes' handlers).

This patch moves the datapaths arrays to a new ovn_datapaths struct,
which wraps the datapaths arrays together with the corresponding
datapaths hmap table, because they are in fact for the same data, and
the array is just an index of the data. The variable for the size of the
array is replaced by a helper function which simply reads the size of
the hmap.

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>