northd: rever ct.inv drop flows
Partially revert the following commit since it introduces a regression
when we want to directly connect to a backend ip from a client outside
the cluster for gw-router-port scenario.

commit e3bc68c3be6967916674119b14fe2bef081ac6ad
Author: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Date:   Mon Mar 20 19:30:13 2023 +0100

    northd: drop ct.inv packets in post snat and lb_aff_learn stages

    Drop ip packets with ct status set to invalid in post snat and
    lb_aff_learn router stages.
    Skip ICMPv{4,6} error messages packet in ct.inv rules in order to
    avoid to introduce too complicated code.

Even if the issue is not directly introduced by the commit above, it is
not easy to fix it without committing all IP traffic to connection
tracking or adding a flow per load-balancer backend.

Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>