tests: update OVS_PAUSE_TEST behavior
User has an option to leave the test environment in error state so that system
can be poked around to get more information. User can enable this option by setting
environment variable OVS_PAUSE_TEST=1. User needs to press CTRL-D to resume the
cleanup operation.

When OVS_PAUSE_TEST=1 and the test succeeds, system is still waiting for CTRL-D
to resume. However, there is no added value to this behavior, as cleanup is already
complete (the only potential added value could be to keep the logs, which can be
achieved using -d option).

This patch causes OVS_PAUSE_TEST=1 to wait for CTRL-D before cleanup only for failed
tests. For successful tests, the test completes as if no OVS_PAUSE_TEST=1.
This new behavior helps in running the same test in loop, with OVS_PAUSE_TEST=1,
and stopping when it fails, and keep environment in error state.
This is useful in trying to reproduce some flaky tests.

Note that the same macro exists in OVS tree. It could be updated there as well
if/once the patch is accepted. ovs-macros are however already slighly different in OVS
and OVN subtrees.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>