tests: Fixed "1 LR with HA distributed router gateway port"
This test is broken since a long time but passed as it used
OVS_WAIT_UNTIL for checking output, which succeeds whatever the output is ...
- Replaced OVS_WAIT_UNTIL by OVS_WAIT_FOR_OUTPUT
- Update table numbers
- Added back hv$i-vif1 port
- Changed slaves to members
- Added back gw2 (hence could remove the XXX)

The following has also been removed:
  OVS_WAIT_UNTIL([as gw2 ovs-ofctl dump-flows br-int table=9 | grep arp_tpa=192.168.0.101 | wc -l], [0], [[0
i.e. checking ARP responder flows in ls_in_arp_rsp as the test expected n such flows for
the lowest priority chassis in the HA group.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>