pinctrl: Simplify packet buffering structure
Simplify the packet buffering process and refactor
the helper functions. The packet buffering functionality
is the same with one exception and the way how we store
buffered packets. Previously it was done through circular
buffer, however considering that the number of packets to
store is 4, convert it to linked list for simplicity.

Two functions in mac-learn are added to public API
temporarily as the functions around packet buffering will
be moved into mac-learn itself.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>