system-tests: Try to load modules only if they weren't loaded before
When the system is busy with multiple test running concurrently
the modprobe might fail with "Device or resource busy", which
will fail the whole test. Instead of always trying to load the module
check if it's already loaded. Also replace the AT_CHECK with OVS_WAIT_UNTIL
which should reduce the chance of failing on initial contention.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>