tests: Use stricter IP match for FORMAT_CT.
The match used in FORMAT_CT matched on "dst=$1". The problem is that
if you passed in an IP address like "10.0.0.1", then this could match on
"10.0.0.1" and "10.0.0.100" for instance. This can allow for unexpected
entries to show up in the output, causing test failures.

With this change, the match is changed to "dst=$1,", which ensures an
exact match on the IP address passed in.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>