controller: Ignore DNS queries with RRs
DNS queries with optional records (RRs), for example, with
cookies for EDNS, are not supported by the OVN resolver.
Trying to reply sometimes results in mangled responses
that clients do not understand.

Instead, just return early when one is present, which
should trigger a negative response and cause clients to
go to the upstream forwarder, hopefully resulting in a
successful query.

Closes issue #192
Signed-off-by: Brian Haley <haleyb.dev@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>