controller: Ignore DNS queries with RRs
DNS queries with optional records (RRs), for example, with
cookies for EDNS, are not supported by the OVN resolver.
Trying to reply will result in mangled responses that
clients do not understand - the ANSWER section will
contain an incorrect option.

Instead, just return early when one is present, which
will trigger a negative response and cause clients to
go to the upstream forwarder, hopefully resulting in a
successful query.

In our testing, the resolver only retries if the
response is correctly formatted, which now happens
with this change.

Closes issue #192
Signed-off-by: Brian Haley <haleyb.dev@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>