ovn-controller: Reduce size of the SB monitor condition.
We don't need to explicitly add port bindings that were already bound
locally.  We implicitly get those because we monitor the datapaths
they're attached to.

When performing an ovn-heater 500-node density-heavy scale test [0], with
conditional monitoring enabled, the unreasonably long poll intervals on
the Southbound database (the ones that took more than one second) are
measured as:

  ---------------------------------------------------------------------
       Count      Min        Max       Median      Mean   95 percentile
  ---------------------------------------------------------------------
        56.0     1010.0     2664.0     1455.5     1544.9     2163.0
        77.0     1015.0     3460.0     1896.0     1858.2     2907.8
        69.0     1010.0     3118.0     1618.0     1688.1     2582.4
  ---------------------------------------------------------------------
       202.0     1010.0     3460.0     1610.0     1713.3     2711.5

Compared to the baseline results (also with conditional monitoring
enabled):
  ---------------------------------------------------------------------
       Count      Min        Max       Median      Mean   95 percentile
  ---------------------------------------------------------------------
       141.0     1003.0    18338.0     1721.0     2625.4     7626.0
       151.0     1019.0    80297.0     1808.0     3410.7     9089.0
       165.0     1007.0    50736.0     2354.0     3067.7     7309.8
  ---------------------------------------------------------------------
       457.0     1003.0    80297.0     2131.0     3044.6     7799.6

We see significant improvement on the server side.  This is explained
by the fact that now the Southbound database server doesn't need to
apply as many conditions as before when filtering individual monitor
contents.

Note: Sub-ports - OVN switch ports with parent_port set - have to be
monitored unconditionally as we cannot efficiently determine their local
datapaths without including all local OVS interfaces in the monitor.
This, however, should not be a huge issue because the majority of ports
are regular VIFs, not sub-ports.

[0] https://github.com/ovn-org/ovn-heater/blob/main/test-scenarios/ocp-500-density-heavy.yml

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2139194
Reported-by: Ilya Maximets <i.maximets@ovn.org>
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>