ovn-northd.c: Omit unused columns in SB_Global.
Connections and SSL columns are not used by ovn-northd. Omit them in
IDL.

This is not a functional problem, but it may hide incremental processing
problems, because when status changes in Connection table, which is
referenced by SB_Global, ovn-northd will receive notifications for
SB_Global change, which triggers recompute. Some tests may pass when
this happens, while it should have failed if the recompute were not
triggered. An example is the test case "testing propagate
Port_Binding.up to NB and OVS", which has passed in most cases but fails
permanently afte this change (and will be fixed in a separate patch).

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>