ovn-northd.at: Fix unstable LSP incremental processing test.
The test case is unstable because there are many factors that can impact
the number of recomputes. For example, when northd updates both NB and
SB, both sides have in-flight transactions, and if the SB notification
comes back before the NB transaction (for updating nb_cfg) completes,
ovn-northd falls back to recompute because idl_txn for NB is not
available, while if the NB notification comes back first it won't
trigger recompute because the nb_cfg columns are omitted for alert.

To avoid the instability of the test case, this patch modifies it to
verify the best case (also the most common scenario), by running the
test multiple (10) times and making sure at least 70% of the time all
the recompute counters match expectation.

(The recompute triggered by in-flight transactions will be improved
later and the success rate in the test can be adjusted accordingly.)

Reported-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>