ovn-northd.at: Fix the LSP incremental processing test case.
The test case intended to ensure there are no more than 3 failures in 10
runs. However, it used "break" to exit the loop whenever a failure is
encountered, end up with at most 1 failure and so the final check for
the failure count will always pass. It should use "continue" instead.

Fixes: 8c30ba13869e ("ovn-northd.at: Fix unstable LSP incremental processing test.")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>