mirror.c: Fix ovn-controller crash when mirror port is deleted from ovs.
If the ovs port used as output_port in mirror is deleted (either by
mistake or intentionally), ovn-controller would crash in the
check_and_update_interface_table() when trying to sync the mirror port
configuration. e.g.:

0  0x0000000000449d6e in check_and_update_interface_table (ovs_mirror=0xcc5110, ovs_mirror=0xcc5110, sb_mirror=0xcc64b0) at ../controller/mirror.c:351
1  sync_ovn_mirror (ovs_mirror_ports=0x7ffcf79b0280, br_int=0xbf5020, ovs_idl_txn=0xe13570, m=0xce86b0) at ../controller/mirror.c:351
2  mirror_run (ovs_idl_txn=0xe13570, ovs_mirror_table=<optimized out>, sb_mirror_table=<optimized out>, br_int=0xbf5020, local_bindings=<optimized out>) at ../controller/mirror.c:174
3  0x0000000000409e8e in main (argc=<optimized out>, argv=<optimized out>) at ../controller/ovn-controller.c:5357

Similarly, it can also crash in delete_ovs_mirror() if the mirror port is
deleted together with src/dst port (when should_delete_ovs_mirror returns
true).

Both cases are captured in the updated test case of this patch.

This patch fixes the problem by verifying the existance of the mirror port, and
creates it if needed.

Fixes: ba8aa26e44cb ("OVN Remote Port Mirroring: controller changes to create ovs mirrors")
Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>