tests: fixed userspace-system tests not properly cleaned up
Usually, when cleaning up a system test we dexecute (for the ovn-vswitchd process):
- OVS_TRAFFIC_VSWITCHD_STOP (i.e. ovs-appctl -t $1 exit --cleanup, wait up to 30 seconds
  and report failure if not properly cleaned up).
- kill_ovs_vswitchd (i.e. if ovs-vsitwhd still running,
  ovs-appctl -t ovs-vswitchd exit --cleanup, wait 1 to 10 seconds and kill ovs-vswitvchd
   if not properly cleaned up).

If we do not execute OVS_TRAFFIC_VSWITCHD_STOP, wait time is much smaller which might result in
- clean up not done properly, ovs-vswitchd killed, ovs-netdev not deleted, and no error reported
- next tests will fail

Using this patch, we properly execute OVS_TRAFFIC_VSWITCHD_STOP, so the delay waiting for
ovs-vswitchd to exit is longer, and, if by any chance this delay is still too short,
the proper test will fail (i.e. the first test failing will be the one unable to clean).

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>