ofctrl-seqno: Do not truncate the last acked value
The requested and acked seqno values are allowed
to be uint64_t, however the values that were added
to the hmap were truncated to uint32_t. This would
lead to loss of information when the value is bigger.

Use uin64_t for the function signatures and for the
hash to prevent truncation.

Reported-at: https://bugzilla.redhat.com/2074019
Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>