binding: fixed ovn-installed not properly removed (migration)
Fixed another potential race condition causing ovn-controller to
not properly remove ovn-install.
Before this patch, ovn-installed was not removed if an hypervisor
was receiving the port_binding->chassis change (to another hv)
within the same idl as the requested-chassis change. This caused
random failures to test "options:multiple requested-chassis for logical port"

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>