northd: Refactor the northd change tracking.
Previous commits made of objdep_mgr to store the lflow references
for ovn_port, ovn_datapath and load balancers.  This patch refactors
the northd change tracking to make it easier to pass the northd
changed data to lflow engine for lflow generation.

In order to incrementally handle the northd changes in the lflow
engine, it is sufficient to provide the below changed data to
lflow engine node
   - changed ovn_ports
   - changed ovn_datapaths
   - changed load balancers.

For every changed resource, lflow northd change handler can get the
objdep_mgr for each of these resources and can regenerate the lflows
easily.  'lflow' changes for ovn_port is already handled.  Upcoming
commits will handle the ovn_datapath and load balancer changes
in lflow engine.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>