ovn-nbctl: Prevent sync exiting early on nb_cfg overflow
We are correctly wrapping nb_cfg over when the value is LLONG_MAX.
However, the sync code is expecting the value to be always raising,
so it would exit early as the current >= next is true for the
wrap back to 0. To detect that use delta between those two values.
The delta threshold is set to INT32_MAX as we don't expect the nb_cfg
to be updated more than that during the sync call.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>