northd: Add incremental processing for NB port groups.
It's similar to the processing we do for address sets.  There's a bit
more mechanics involved due to the fact that we need to split NB port
groups per datapath.

We currently only partially implement incremental processing of
port_group changes in the lflow node.  That is, we deal with the case
when the sets of "switches per port group" doesn't change.  In that
specific case ACL lflows don't need to be reprocessed.

In a synthetic benchmark that created (in this order):
- 500 switches
- 2000 port groups
- 4 ACLs per port group
- 10000 ports distributed equally between the switches and port groups

we measured the following ovn-northd CPU usage:

  +-------------------------+------------+--------------------+
  | Incremental processing? | --wait=sb? | northd avg cpu (%) |
  +-------------------------+------------+--------------------+
  |           N             |     Y      |        84.2        |
  +-------------------------+------------+--------------------+
  |           Y             |     Y      |        41.5        |
  +-------------------------+------------+--------------------+
  |           N             |     N      |        93.2        |
  +-------------------------+------------+--------------------+
  |           Y             |     N      |        53.6        |
  +-------------------------+------------+--------------------+

where '--wait=sb' set to 'Y'  means the benchmark was waiting for the
port and port group operations to be propagated to the Southbound DB
before continuing to the next operation.

Reported-at: https://bugzilla.redhat.com/show_bug.cgi?id=2228162
Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>