ci, tests: Remove parallelization permutation
Single test needs to run 4 times to check all
the permutations, this is taking a lot of CI time
and job space.

Remove the parallel permutation and leave parallelization
enabled for all tests, as this use case is more complex.
This allows us to greatly reduce the number of test cases
(by half) and we can also remove 6 jobs from the CI pipeline.

The time reduction is very noticeable going down from ~30 min
to ~20 min.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>