northd: Support an option to ignore chassis features.
Add option ignore_chassis_features for northd to bypass the support
status of features on each chassis and to directly implement the latest
features.

This is particularly useful for users who follow the suggested upgrade
order that upgrades ovn-controllers before ovn-northd, and want to
safeguard the operation of northd from being adversely affected by a
mismatched configuration of a chassis. This not only avoids feature
degradation but can also avoid risks of any broken feature in backward
compatible mode. An example of such impact is discussed at [0].

[0] https://mail.openvswitch.org/pipermail/ovs-dev/2023-September/407756.html

Signed-off-by: Han Zhou <hzhou@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>