ci, tests: Use parallelization permutations for few jobs
Single test needs to run 4 times to check all
the permutations, this is taking a lot of CI time
and job space.

Remove the parallel permutation and leave parallelization
enabled for all tests, as this use case is more complex.
Only exception are 10 tests in ovn-northd.at that still run
parallelization permutations. This should be enough to cover
the code with parallelization disabled.

This allows us to greatly reduce the number of test cases
(almost by half) and we can also remove 6 jobs from the CI
pipeline. The time reduction is very noticeable going down
from ~30 min to ~20 min.

Signed-off-by: Ales Musil <amusil@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>