controller: throttle port claim attempts from if-status
A recent commit (4dc4bc7f) throttled the port claim attemps from the binding module
when multiple chassis were fighting for the same port.
However, the if-status module was also sometimes fighting for the same ports.
Fix this by having the if-status module postponing claims in a similar way as the binding module.

Signed-off-by: Xavier Simonart <xsimonar@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>