tests: Use fmt_pkt in 3 HVs, 1 LS, 3 lports/HV.
This only uses fmt_pkt in test_arp(). test_packet() purposely sends
packets with odd and/or malformed ethertypes, which makes it a poor fit
for fmt_pkt(). We could use some sort of raw packet method of scapy, but
that seems like a roundabout way of getting what we already have in the
test.

Execution time: 10.313s
Execution time on "main" branch: 8.575s

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>