tests: Add missing check for scapy.
The "IPv6 ND Router Solicitation responder" test was altered to test
more cases. When the test was updated, the author updated the test to
use fmt_pkt instead of constructing the packets by hand. Unfortunately,
this change did not include the directive to skip the test if scapy is
not installed. This commit adds the check so that systems that do not
have scapy installed can still run tests appropriately.

Fixes: 7b70e07ce ("pinctrl: Reply with correct destination for ICMPv6 RA packets")
Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>