northd: Add a new engine 'lr-lb-nat-data' to manage routers' lb data.
This new engine now maintains the load balancer and NAT data of a
logical router which was earlier part of northd engine node data.
The main inputs to this engine are:
   - northd node
   - lr-nat node

A record for each logical router is maintained in the 'lr_lb_nats'
hmap table and this record
   - stores the lb related data
   - embeds the 'lr-nat' record.

This engine node becomes an input to 'lflow' node.

Signed-off-by: Numan Siddique <numans@ovn.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>