check_acl_log: Format log printing according to E231 flake8 rule.
Without this, when using Python 3.12 and flake8 5.0.4, the following
errors are flagged:
  tests/check_acl_log.py:97:25: E231 missing whitespace after ':'
  tests/check_acl_log.py:102:71: E231 missing whitespace after ':'

While this seems a bit too restrictive from flake8 we can easily work
around it by just not using f-strings unless necessary.

This unblocks GitHub CI runs on branches <= 23.06.  These started
failing since the Python version we get in CI (in the Ubuntu runner)
changed to CPython (3.12.0).

Signed-off-by: Dumitru Ceara <dceara@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>