tests: Use fmt_pkt in 3 HVs, 3 LS, 3 lports/LS.
This test is slower than it used to be since some of the test_ip calls
that were backgrounded (with '&') cannot do this anymore. This is
because there is a race condition with starting the scapy server when
the calls are backgrounded.

Execution time: 36.471s
Execution time on "main" branch: 13.913s

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>